+++ tests/test_ast_nodes_core.py
@@ -1,20 +1,13 @@
-"""Complete AST Nodes Tests for Issue #491 Phase 3
+"""AST Nodes Core Tests
 
-Comprehensive tests for ast_nodes module to achieve high coverage.
+Split from test_ast_nodes_complete.py for 300-line limit compliance.
+Comprehensive tests for Node class core functionality.
 Current coverage: node.py 52% → Target: 90%+
 """
 
 import pytest
 
-from kumihan_formatter.core.ast_nodes import Node, NodeBuilder
-from kumihan_formatter.core.ast_nodes.factories import (
-    div_box,
-    emphasis,
-    heading,
-    highlight,
-    paragraph,
-    strong,
-)
+from kumihan_formatter.core.ast_nodes import Node
 
 
 class TestNodeCompleteCoverage:
@@ -330,299 +323,3 @@ class TestNodeCompleteCoverage:
         # Repr method
         repr_str = repr(node)
         assert "Node" in repr_str
-
-
-class TestNodeBuilderCompleteCoverage:
-    """Complete coverage tests for NodeBuilder class"""
-
-    def test_node_builder_all_methods(self):
-        """Test all NodeBuilder methods"""
-        builder = NodeBuilder("section")
-
-        # Method chaining
-        node = (
-            builder.content("Test content")
-            .css_class("section-class")
-            .id("section-1")
-            .style("color: red;")
-            .attribute("data-test", "value")
-            .attribute("role", "main")
-            .build()
-        )
-
-        assert node.type == "section"
-        assert node.content == "Test content"
-        assert node.get_attribute("class") == "section-class"
-        assert node.get_attribute("id") == "section-1"
-        assert node.get_attribute("style") == "color: red;"
-        assert node.get_attribute("data-test") == "value"
-        assert node.get_attribute("role") == "main"
-
-    def test_node_builder_different_content_types(self):
-        """Test NodeBuilder with different content types"""
-        # String content
-        builder = NodeBuilder("p")
-        node = builder.content("String content").build()
-        assert node.content == "String content"
-
-        # List content
-        builder = NodeBuilder("div")
-        list_content = ["item1", "item2"]
-        node = builder.content(list_content).build()
-        assert node.content == list_content
-
-        # Node content
-        child_node = Node("span", "child")
-        builder = NodeBuilder("div")
-        node = builder.content(child_node).build()
-        assert node.content == child_node
-
-        # None content
-        builder = NodeBuilder("br")
-        node = builder.content(None).build()
-        assert node.content is None
-
-    def test_node_builder_attribute_overwrites(self):
-        """Test NodeBuilder attribute overwrite behavior"""
-        builder = NodeBuilder("div")
-
-        # Overwrite same attribute
-        node = (
-            builder.css_class("first-class")
-            .css_class("second-class")  # Should overwrite
-            .build()
-        )
-
-        assert node.get_attribute("class") == "second-class"
-
-        # Overwrite with attribute method
-        builder = NodeBuilder("span")
-        node = (
-            builder.id("first-id")
-            .attribute("id", "second-id")  # Should overwrite
-            .build()
-        )
-
-        assert node.get_attribute("id") == "second-id"
-
-    def test_node_builder_edge_cases(self):
-        """Test NodeBuilder edge cases"""
-        # Empty type
-        builder = NodeBuilder("")
-        node = builder.content("content").build()
-        assert node.type == ""
-
-        # Very long type
-        long_type = "very_long_element_type_name_for_testing"
-        builder = NodeBuilder(long_type)
-        node = builder.build()
-        assert node.type == long_type
-
-        # Build without setting anything
-        builder = NodeBuilder("div")
-        node = builder.build()
-        assert node.type == "div"
-        assert node.content is None
-        assert node.attributes == {}
-
-        # Multiple builds from same builder
-        builder = NodeBuilder("p").content("text")
-        node1 = builder.build()
-        node2 = builder.build()
-        assert node1.content == node2.content
-        assert node1 is not node2  # Different objects
-
-
-class TestNodeFactoriesComplete:
-    """Complete coverage tests for Node factory functions"""
-
-    def test_paragraph_factory(self):
-        """Test paragraph factory function"""
-        # Simple paragraph
-        p = paragraph("Text content")
-        assert p.type == "p"
-        assert p.content == "Text content"
-
-        # Paragraph with attributes
-        p_with_class = paragraph("Styled text", {"class": "intro"})
-        assert p_with_class.get_attribute("class") == "intro"
-
-    def test_heading_factory(self):
-        """Test heading factory function"""
-        # Different heading levels
-        for level in [1, 2, 3, 4, 5]:
-            h = heading(level, "Heading text")
-            assert h.type == f"h{level}"
-            assert h.content == "Heading text"
-
-        # Heading with attributes
-        h_with_id = heading(1, "Title", {"id": "main-title"})
-        assert h_with_id.get_attribute("id") == "main-title"
-
-        # Invalid level handling
-        try:
-            h_invalid = heading(0, "Invalid")
-            # Should handle gracefully or raise appropriate error
-        except (ValueError, AssertionError):
-            pass  # Expected behavior
-
-    def test_container_factories(self):
-        """Test container factory functions"""
-        # Div box factory
-        d = div_box("Div content")
-        assert d.type == "div"
-        assert d.content == "Div content"
-        assert d.get_attribute("class") == "box"
-
-        # Highlight div factory
-        h = highlight("Highlighted content")
-        assert h.type == "div"
-        assert h.content == "Highlighted content"
-        assert h.get_attribute("class") == "highlight"
-
-        # Highlight with color
-        h_colored = highlight("Colored content", "yellow")
-        assert "background-color:yellow" in h_colored.get_attribute("style", "")
-
-    def test_formatting_factories(self):
-        """Test formatting factory functions"""
-        # Strong (bold)
-        strong_node = strong("Bold text")
-        assert strong_node.type == "strong"
-        assert strong_node.content == "Bold text"
-
-        # Emphasis (italic)
-        em_node = emphasis("Italic text")
-        assert em_node.type == "em"
-        assert em_node.content == "Italic text"
-
-    def test_manual_node_creation(self):
-        """Test manual node creation for missing factories"""
-        # List item (manual creation)
-        li = Node("li", "Item content")
-        assert li.type == "li"
-        assert li.content == "Item content"
-
-        # Unordered list (manual creation)
-        item1 = Node("li", "Item 1")
-        item2 = Node("li", "Item 2")
-        ul = Node("ul", [item1, item2])
-        assert ul.type == "ul"
-        assert len(ul.content) == 2
-
-        # Ordered list (manual creation)
-        ol = Node("ol", [item1, item2])
-        assert ol.type == "ol"
-        assert len(ol.content) == 2
-
-        # Error node (manual creation)
-        error = Node("error", "Error message")
-        assert error.type == "error"
-        assert "Error message" in error.content
-
-
-class TestNodeUtilitiesComplete:
-    """Complete coverage tests for Node utilities"""
-
-    def test_node_tree_traversal(self):
-        """Test node tree traversal utilities"""
-        # Create tree structure
-        root = div("Root")
-        child1 = paragraph("Child 1")
-        child2 = div("Child 2")
-        grandchild = span("Grandchild")
-        child2.add_child(grandchild)
-        root.content = [child1, child2]
-
-        # Basic tree structure validation
-        assert isinstance(root.content, list)
-        assert len(root.content) == 2
-        assert root.content[0] == child1
-        assert root.content[1] == child2
-
-        # Child content validation
-        assert isinstance(child2.content, list)
-        assert len(child2.content) == 1
-        assert child2.content[0] == grandchild
-
-        # Attribute testing
-        child1.add_attribute("class", "special")
-        assert child1.get_attribute("class") == "special"
-
-    def test_node_transformation_utilities(self):
-        """Test node transformation utilities"""
-        # Manual type conversion
-        p_node = paragraph("Text")
-        div_node = Node("div", p_node.content)
-        div_node.attributes = p_node.attributes.copy()
-        assert div_node.type == "div"
-        assert div_node.content == "Text"
-
-        # Manual wrapping
-        text_node = Node("span", "Text")
-        wrapped = Node("div", [text_node])
-        wrapped.add_attribute("class", "wrapper")
-        assert wrapped.type == "div"
-        assert wrapped.get_attribute("class") == "wrapper"
-        assert wrapped.content == [text_node]
-
-        # Manual unwrapping
-        container = Node("div", [Node("span", "Content")])
-        unwrapped = container.content
-        assert len(unwrapped) == 1
-        assert unwrapped[0].type == "span"
-        assert unwrapped[0].content == "Content"
-
-    def test_node_validation_utilities(self):
-        """Test node validation utilities"""
-        # Basic node validation
-        valid_node = paragraph("Valid content")
-        assert valid_node.type == "p"
-        assert valid_node.content == "Valid content"
-
-        # Edge case node
-        edge_node = Node("", None)
-        assert edge_node.type == ""
-        assert edge_node.content is None
-
-        # Check attributes existence
-        node_with_attrs = Node("div", "Content")
-        node_with_attrs.add_attribute("id", "required-id")
-
-        # Attribute validation
-        assert node_with_attrs.has_attribute("id")
-        assert not node_with_attrs.has_attribute("class")
-        assert node_with_attrs.get_attribute("id") == "required-id"
-
-    def test_node_serialization_utilities(self):
-        """Test node serialization utilities"""
-        # Manual dictionary creation
-        node = paragraph("Content")
-        node.add_attribute("class", "test")
-
-        node_dict = {
-            "type": node.type,
-            "content": node.content,
-            "attributes": node.attributes,
-        }
-        assert node_dict["type"] == "p"
-        assert node_dict["content"] == "Content"
-        assert node_dict["attributes"]["class"] == "test"
-
-        # Manual reconstruction
-        reconstructed = Node(
-            node_dict["type"], node_dict["content"], node_dict["attributes"]
-        )
-        assert reconstructed.type == node.type
-        assert reconstructed.content == node.content
-        assert reconstructed.attributes == node.attributes
-
-        # JSON compatibility test
-        import json
-
-        json_str = json.dumps(node_dict)
-        parsed_dict = json.loads(json_str)
-        json_node = Node(
-            parsed_dict["type"], parsed_dict["content"], parsed_dict["attributes"]
-        )
-        assert json_node.type == node.type
