+++ tests/test_filesystem_and_data_utilities.py
@@ -1,0 +1,195 @@
+"""Filesystem and Data Conversion Utilities Tests
+
+Split from test_utilities_complete.py for 300-line limit compliance.
+Tests for file system operations and data conversion utilities.
+"""
+
+import tempfile
+from pathlib import Path
+
+import pytest
+
+
+class TestFileSystemUtilities:
+    """Test file system utilities"""
+
+    def test_file_system_operations(self):
+        """Test file system operations"""
+        try:
+            from kumihan_formatter.core.utilities.file_system import (
+                copy_with_metadata,
+                ensure_directory,
+                get_file_info,
+                safe_remove,
+            )
+
+            with tempfile.TemporaryDirectory() as temp_dir:
+                temp_path = Path(temp_dir)
+
+                # Test ensure_directory
+                try:
+                    new_dir = temp_path / "test_dir"
+                    ensure_directory(new_dir)
+                    assert new_dir.exists()
+                    assert new_dir.is_dir()
+                except:
+                    pass
+
+                # Test get_file_info
+                test_file = temp_path / "test.txt"
+                test_file.write_text("Test content")
+
+                try:
+                    info = get_file_info(test_file)
+                    assert isinstance(info, dict)
+                    assert "size" in info
+                    assert "modified" in info
+                except:
+                    pass
+
+                # Test safe_remove
+                try:
+                    safe_remove(test_file)
+                    assert not test_file.exists()
+                except:
+                    pass
+
+        except ImportError:
+            # File system utilities may not be available
+            pass
+
+    def test_path_utilities(self):
+        """Test path utilities"""
+        try:
+            from kumihan_formatter.core.utilities.file_system import (
+                is_safe_path,
+                normalize_path,
+                resolve_path,
+            )
+
+            # Test normalize_path
+            try:
+                normalized = normalize_path("./test/../file.txt")
+                assert isinstance(normalized, (str, Path))
+            except:
+                pass
+
+            # Test resolve_path
+            try:
+                resolved = resolve_path("~/test.txt")
+                assert isinstance(resolved, (str, Path))
+            except:
+                pass
+
+            # Test is_safe_path
+            try:
+                assert is_safe_path("/tmp/test.txt") in [True, False]
+                assert is_safe_path("../../../etc/passwd") is False
+            except:
+                pass
+
+        except ImportError:
+            pass
+
+
+class TestDataConverters:
+    """Test data converter utilities"""
+
+    def test_converters_functionality(self):
+        """Test converter functionality"""
+        try:
+            from kumihan_formatter.core.utilities.converters import (
+                from_json,
+                from_yaml,
+                to_dict,
+                to_json,
+                to_yaml,
+            )
+
+            # Test data
+            test_data = {
+                "name": "test",
+                "value": 123,
+                "nested": {"key": "value"},
+                "list": [1, 2, 3],
+            }
+
+            # Test to_json/from_json
+            try:
+                json_str = to_json(test_data)
+                assert isinstance(json_str, str)
+
+                parsed = from_json(json_str)
+                assert parsed == test_data
+            except:
+                pass
+
+            # Test to_dict
+            try:
+
+                class TestObj:
+                    def __init__(self):
+                        self.attr1 = "value1"
+                        self.attr2 = "value2"
+
+                obj = TestObj()
+                dict_result = to_dict(obj)
+                assert isinstance(dict_result, dict)
+                assert "attr1" in dict_result
+            except:
+                pass
+
+        except ImportError:
+            pass
+
+    def test_type_converters(self):
+        """Test type converters"""
+        try:
+            from kumihan_formatter.core.utilities.converters import (
+                to_bool,
+                to_float,
+                to_int,
+                to_list,
+            )
+
+            # Test to_bool
+            bool_tests = [
+                ("true", True),
+                ("false", False),
+                ("1", True),
+                ("0", False),
+                ("yes", True),
+                ("no", False),
+            ]
+
+            for value, expected in bool_tests:
+                try:
+                    result = to_bool(value)
+                    assert result == expected
+                except:
+                    pass
+
+            # Test to_int
+            try:
+                assert to_int("123") == 123
+                assert to_int("123.45") == 123
+                assert to_int("invalid", default=0) == 0
+            except:
+                pass
+
+            # Test to_float
+            try:
+                assert to_float("123.45") == 123.45
+                assert to_float("123") == 123.0
+            except:
+                pass
+
+            # Test to_list
+            try:
+                assert to_list("a,b,c") == ["a", "b", "c"]
+                assert to_list([1, 2, 3]) == [1, 2, 3]
+            except:
+                pass
+
+        except ImportError:
+            pass
