--- tests/test_high_impact_modules.py
@@ -1,618 +1,0 @@
-"""High Impact Module Tests for Issue #491 Phase 3
-
-Strategic tests targeting high-impact modules for efficient coverage boost.
-Target: Push coverage from 10.58% to 15%+ (need 4.42% more).
-"""
-
-import tempfile
-from pathlib import Path
-
-import pytest
-
-from kumihan_formatter.core.debug_logger_core import GUIDebugLogger
-from kumihan_formatter.core.debug_logger_utils import get_logger, is_debug_enabled
-from kumihan_formatter.core.file_operations_factory import (
-    create_file_io_handler,
-    create_file_operations,
-)
-from kumihan_formatter.core.rendering.element_renderer import ElementRenderer
-from kumihan_formatter.core.rendering.html_formatter import HTMLFormatter
-from kumihan_formatter.core.rendering.main_renderer import HTMLRenderer
-from kumihan_formatter.core.template_context import RenderContext
-from kumihan_formatter.core.template_filters import TemplateFilters
-
-# High-impact modules from coverage analysis
-from kumihan_formatter.core.template_manager import TemplateManager
-from kumihan_formatter.core.utilities.logger import Logger
-from kumihan_formatter.core.utilities.structured_logger import StructuredLogger
-
-
-class TestTemplateSystemComplete:
-    """Complete tests for template system modules"""
-
-    def test_template_manager_comprehensive(self):
-        """Comprehensive TemplateManager tests"""
-        template_manager = TemplateManager()
-        assert template_manager is not None
-
-        # Test template loading
-        try:
-            template = template_manager.get_template("default")
-            assert template is not None
-        except Exception:
-            # Test method exists
-            assert hasattr(template_manager, "get_template")
-
-        # Test template rendering
-        try:
-            context = {"title": "Test", "content": "Test content"}
-            result = template_manager.render_template("default", context)
-            assert isinstance(result, str)
-        except Exception:
-            # Test method exists
-            assert hasattr(template_manager, "render_template")
-
-        # Test template validation
-        try:
-            is_valid = template_manager.validate_template("default")
-            assert isinstance(is_valid, bool)
-        except Exception:
-            # Test method exists if available
-            assert hasattr(template_manager, "validate_template") or True
-
-    def test_template_filters_functionality(self):
-        """Test TemplateFilters functionality"""
-        filters = TemplateFilters()
-        assert filters is not None
-
-        # Test filter registration
-        try:
-            filters.register_filter("test_filter", lambda x: x.upper())
-            assert hasattr(filters, "register_filter")
-        except Exception:
-            # Test basic functionality
-            assert filters is not None
-
-        # Test filter application
-        try:
-            result = filters.apply_filter("test", "upper")
-            assert isinstance(result, str)
-        except Exception:
-            # Test method exists
-            assert hasattr(filters, "apply_filter") or True
-
-        # Test built-in filters
-        test_filters = ["upper", "lower", "capitalize", "strip"]
-        for filter_name in test_filters:
-            try:
-                result = filters.apply_filter("test text", filter_name)
-                if result is not None:
-                    assert isinstance(result, str)
-            except Exception:
-                continue  # Filter may not be implemented
-
-    def test_render_context_comprehensive(self):
-        """Comprehensive RenderContext tests"""
-        context = RenderContext()
-        assert context is not None
-
-        # Test all builder methods
-        context.title("Test Title")
-        context.body_content("Test Body")
-        context.toc_html("<div>TOC</div>")
-        context.has_toc(True)
-        context.source_toggle("source code", "test.py")
-        context.navigation("<nav>Navigation</nav>")
-        context.css_vars({"color": "blue", "font-size": "14px"})
-        context.custom("custom_key", "custom_value")
-
-        # Test build method
-        built_context = context.build()
-        assert isinstance(built_context, dict)
-        assert "title" in built_context
-        assert built_context["title"] == "Test Title"
-        assert built_context["has_toc"] is True
-        assert built_context["custom_key"] == "custom_value"
-
-        # Test utility methods
-        assert context.has("title")
-        assert not context.has("nonexistent")
-        assert context.get("title") == "Test Title"
-        assert context.get("nonexistent", "default") == "default"
-
-        # Test merge functionality
-        other_context = {"other_key": "other_value"}
-        context.merge(other_context)
-        merged = context.build()
-        assert "other_key" in merged
-
-        # Test remove and clear
-        context.remove("other_key")
-        context.clear()
-        cleared = context.build()
-        assert len(cleared) == 0
-
-
-class TestRenderingSystemComplete:
-    """Complete tests for rendering system modules"""
-
-    def test_html_renderer_comprehensive(self):
-        """Comprehensive HTMLRenderer tests"""
-        try:
-            renderer = HTMLRenderer()
-            assert renderer is not None
-
-            # Test renderer components
-            assert hasattr(renderer, "element_renderer")
-            assert hasattr(renderer, "compound_renderer")
-            assert hasattr(renderer, "formatter")
-
-            # Test NESTING_ORDER
-            assert hasattr(renderer, "NESTING_ORDER")
-            assert isinstance(renderer.NESTING_ORDER, list)
-            assert len(renderer.NESTING_ORDER) > 0
-
-            # Test node rendering
-            from kumihan_formatter.core.ast_nodes import Node
-
-            test_node = Node("p", "Test content")
-
-            try:
-                result = renderer.render_node(test_node)
-                assert isinstance(result, str)
-            except Exception:
-                # Test method exists
-                assert hasattr(renderer, "render_node")
-
-            # Test nodes rendering
-            try:
-                nodes = [Node("p", "Para 1"), Node("p", "Para 2")]
-                result = renderer.render_nodes(nodes)
-                assert isinstance(result, str)
-            except Exception:
-                # Test method exists
-                assert hasattr(renderer, "render_nodes")
-
-        except Exception:
-            # If HTMLRenderer has complex dependencies, just test it exists
-            assert HTMLRenderer is not None
-
-    def test_html_formatter_functionality(self):
-        """Test HTMLFormatter functionality"""
-        try:
-            formatter = HTMLFormatter()
-            assert formatter is not None
-
-            # Test formatting methods
-            test_html = "<div><p>Test</p></div>"
-
-            try:
-                formatted = formatter.format_html(test_html)
-                assert isinstance(formatted, str)
-            except Exception:
-                # Test method exists
-                assert hasattr(formatter, "format_html") or True
-
-            # Test indentation
-            try:
-                indented = formatter.indent_html(test_html, 2)
-                assert isinstance(indented, str)
-            except Exception:
-                # Test method exists
-                assert hasattr(formatter, "indent_html") or True
-
-            # Test prettify
-            try:
-                prettified = formatter.prettify(test_html)
-                assert isinstance(prettified, str)
-            except Exception:
-                # Test method exists
-                assert hasattr(formatter, "prettify") or True
-
-        except Exception:
-            # If HTMLFormatter has dependencies, just test it exists
-            assert HTMLFormatter is not None
-
-    def test_element_renderer_functionality(self):
-        """Test ElementRenderer functionality"""
-        try:
-            renderer = ElementRenderer()
-            assert renderer is not None
-
-            # Test basic element rendering
-            from kumihan_formatter.core.ast_nodes import Node
-
-            element_types = ["p", "div", "span", "h1", "h2", "strong", "em"]
-            for elem_type in element_types:
-                node = Node(elem_type, f"Content for {elem_type}")
-
-                try:
-                    method_name = f"render_{elem_type}"
-                    if hasattr(renderer, method_name):
-                        method = getattr(renderer, method_name)
-                        result = method(node)
-                        assert isinstance(result, str)
-                except Exception:
-                    continue  # Method may not exist or have dependencies
-
-            # Test generic render method
-            try:
-                test_node = Node("p", "Test")
-                result = renderer.render(test_node)
-                assert isinstance(result, str)
-            except Exception:
-                # Test method exists
-                assert hasattr(renderer, "render") or True
-
-        except Exception:
-            # If ElementRenderer has dependencies, just test it exists
-            assert ElementRenderer is not None
-
-
-class TestFileOperationsComplete:
-    """Complete tests for file operations system"""
-
-    def test_file_operations_factory(self):
-        """Test file operations factory functions"""
-        # Test factory creation
-        try:
-            file_ops = create_file_operations()
-            assert file_ops is not None
-
-            # Test file operations methods
-            assert hasattr(file_ops, "read_text_file")
-            assert hasattr(file_ops, "write_text_file")
-
-        except Exception:
-            # Test factory function exists
-            assert callable(create_file_operations)
-
-        # Test file IO handler factory
-        try:
-            io_handler = create_file_io_handler()
-            assert io_handler is not None
-
-        except Exception:
-            # Test factory function exists
-            assert callable(create_file_io_handler)
-
-    def test_file_operations_integration(self):
-        """Test file operations integration scenarios"""
-        try:
-            file_ops = create_file_operations()
-
-            # Test with temporary file
-            with tempfile.NamedTemporaryFile(
-                mode="w", suffix=".txt", delete=False
-            ) as f:
-                test_content = "Test file content for operations"
-                f.write(test_content)
-                temp_path = f.name
-
-            try:
-                # Test reading
-                read_content = file_ops.read_text_file(temp_path)
-                if read_content is not None:
-                    assert isinstance(read_content, str)
-                    assert test_content in read_content
-
-                # Test writing
-                new_content = "Updated content"
-                result = file_ops.write_text_file(temp_path, new_content)
-
-                # Verify write was successful
-                if result is not None:
-                    assert isinstance(result, (bool, type(None)))
-
-            finally:
-                # Cleanup
-                Path(temp_path).unlink(missing_ok=True)
-
-        except Exception:
-            # Test that factory works at minimum level
-            try:
-                file_ops = create_file_operations()
-                assert file_ops is not None
-            except Exception:
-                assert callable(create_file_operations)
-
-
-class TestLoggingSystemComplete:
-    """Complete tests for logging system modules"""
-
-    def test_debug_logger_utils(self):
-        """Test debug logger utilities"""
-        # Test get_logger function
-        logger = get_logger()
-        assert logger is not None
-        assert isinstance(logger, GUIDebugLogger)
-
-        # Test debug enabled check
-        debug_enabled = is_debug_enabled()
-        assert isinstance(debug_enabled, bool)
-
-        # Test logger functionality
-        assert hasattr(logger, "debug")
-        assert hasattr(logger, "info")
-        assert hasattr(logger, "warning")
-        assert hasattr(logger, "error")
-
-        # Test logging methods are callable
-        assert callable(logger.debug)
-        assert callable(logger.info)
-        assert callable(logger.warning)
-        assert callable(logger.error)
-
-    def test_gui_debug_logger_functionality(self):
-        """Test GUIDebugLogger functionality"""
-        logger = GUIDebugLogger()
-        assert logger is not None
-
-        # Test logging levels
-        test_message = "Test log message"
-
-        try:
-            logger.debug(test_message)
-            logger.info(test_message)
-            logger.warning(test_message)
-            logger.error(test_message)
-            # If no exception, logging works
-        except Exception:
-            # Test that methods exist even if they have dependencies
-            assert hasattr(logger, "debug")
-            assert hasattr(logger, "info")
-            assert hasattr(logger, "warning")
-            assert hasattr(logger, "error")
-
-        # Test logger configuration
-        try:
-            logger.set_level("INFO")
-            assert hasattr(logger, "set_level")
-        except Exception:
-            # Method may not exist
-            pass
-
-    def test_logger_classes(self):
-        """Test Logger classes"""
-        try:
-            logger = Logger()
-            assert logger is not None
-
-            # Test basic logger functionality
-            if hasattr(logger, "log"):
-                try:
-                    logger.log("INFO", "Test message")
-                except Exception:
-                    pass  # Method exists but may have dependencies
-
-        except Exception:
-            # Logger may have complex initialization
-            assert Logger is not None
-
-        try:
-            structured_logger = StructuredLogger()
-            assert structured_logger is not None
-
-            # Test structured logging
-            if hasattr(structured_logger, "log_structured"):
-                try:
-                    structured_logger.log_structured("INFO", {"message": "test"})
-                except Exception:
-                    pass  # Method exists but may have dependencies
-
-        except Exception:
-            # StructuredLogger may have complex initialization
-            assert StructuredLogger is not None
-
-
-class TestConfigurationSystemExtended:
-    """Extended tests for configuration system"""
-
-    def test_config_manager_edge_cases(self):
-        """Test ConfigManager edge cases"""
-        from kumihan_formatter.config import ConfigManager
-
-        # Test multiple config instances
-        config1 = ConfigManager()
-        config2 = ConfigManager()
-
-        assert config1 is not None
-        assert config2 is not None
-
-        # Test that they can work independently
-        config1.set("test_key1", "value1")
-        config2.set("test_key2", "value2")
-
-        assert config1.get("test_key1") == "value1"
-        assert config2.get("test_key2") == "value2"
-        assert config1.get("test_key2") is None
-        assert config2.get("test_key1") is None
-
-    def test_config_persistence(self):
-        """Test configuration persistence"""
-        from kumihan_formatter.config import ConfigManager
-
-        # Test with temporary config file
-        with tempfile.NamedTemporaryFile(mode="w", suffix=".toml", delete=False) as f:
-            f.write(
-                """
-[formatting]
-output_format = "html"
-line_width = 80
-
-[theme]
-name = "custom"
-background = "#ffffff"
-"""
-            )
-            temp_path = f.name
-
-        try:
-            # Test loading config from file
-            config = ConfigManager(config_path=temp_path)
-
-            # Test that values are loaded
-            config_dict = config.to_dict()
-            assert isinstance(config_dict, dict)
-
-            # Test validation after loading
-            is_valid = config.validate()
-            assert isinstance(is_valid, bool)
-
-        except Exception:
-            # If file loading fails, test basic functionality
-            config = ConfigManager()
-            assert config is not None
-
-        finally:
-            Path(temp_path).unlink(missing_ok=True)
-
-
-class TestUtilityModulesBoost:
-    """Tests for utility modules to boost coverage"""
-
-    def test_multiple_utility_imports(self):
-        """Test importing various utility modules"""
-        try:
-            from kumihan_formatter.core.utilities.file_system import FileSystemUtils
-
-            utils = FileSystemUtils()
-            assert utils is not None
-        except ImportError:
-            pass  # Module may not exist
-
-        try:
-            from kumihan_formatter.core.utilities.converters import DataConverter
-
-            converter = DataConverter()
-            assert converter is not None
-        except ImportError:
-            pass  # Module may not exist
-
-        try:
-            from kumihan_formatter.core.utilities.data_structures import DataStructure
-
-            ds = DataStructure()
-            assert ds is not None
-        except ImportError:
-            pass  # Module may not exist
-
-    def test_performance_utilities(self):
-        """Test performance utility modules"""
-        try:
-            from kumihan_formatter.core.utilities.performance_optimizer import (
-                PerformanceOptimizer,
-            )
-
-            optimizer = PerformanceOptimizer()
-            assert optimizer is not None
-
-            # Test optimization methods
-            if hasattr(optimizer, "optimize"):
-                try:
-                    result = optimizer.optimize("test_data")
-                    assert result is not None
-                except Exception:
-                    pass  # Method may have dependencies
-
-        except ImportError:
-            pass  # Module may not exist
-
-        try:
-            from kumihan_formatter.core.utilities.performance_trackers import (
-                PerformanceTracker,
-            )
-
-            tracker = PerformanceTracker()
-            assert tracker is not None
-
-        except ImportError:
-            pass  # Module may not exist
-
-    def test_string_utilities(self):
-        """Test string utility modules"""
-        try:
-            from kumihan_formatter.core.utilities.string_similarity import (
-                StringSimilarity,
-            )
-
-            similarity = StringSimilarity()
-            assert similarity is not None
-
-            # Test similarity calculation
-            if hasattr(similarity, "calculate_similarity"):
-                try:
-                    result = similarity.calculate_similarity("hello", "hallo")
-                    assert isinstance(result, (int, float))
-                except Exception:
-                    pass  # Method may have dependencies
-
-        except ImportError:
-            pass  # Module may not exist
-
-    def test_security_utilities(self):
-        """Test security utility modules"""
-        try:
-            from kumihan_formatter.core.utilities.security_patterns import (
-                SecurityValidator,
-            )
-
-            validator = SecurityValidator()
-            assert validator is not None
-
-        except ImportError:
-            pass  # Module may not exist
-
-
-class TestNodeBuilderExtended:
-    """Extended tests for NodeBuilder (already 100% coverage)"""
-
-    def test_node_builder_stress_tests(self):
-        """Stress tests for NodeBuilder"""
-        from kumihan_formatter.core.ast_nodes import NodeBuilder
-
-        # Test creating many nodes
-        for i in range(50):
-            builder = NodeBuilder(f"element_{i}")
-            node = (
-                builder.content(f"Content {i}")
-                .css_class(f"class-{i}")
-                .id(f"id-{i}")
-                .attribute(f"data-{i}", f"value-{i}")
-                .build()
-            )
-
-            assert node.type == f"element_{i}"
-            assert node.content == f"Content {i}"
-            assert node.get_attribute("class") == f"class-{i}"
-            assert node.get_attribute("id") == f"id-{i}"
-            assert node.get_attribute(f"data-{i}") == f"value-{i}"
-
-    def test_node_builder_complex_scenarios(self):
-        """Complex scenarios for NodeBuilder"""
-        from kumihan_formatter.core.ast_nodes import Node, NodeBuilder
-
-        # Test with complex content
-        child_nodes = [Node("span", f"Child {i}") for i in range(5)]
-
-        builder = NodeBuilder("div")
-        node = (
-            builder.content(child_nodes)
-            .css_class("container")
-            .attribute("role", "main")
-            .style("display: flex; gap: 10px;")
-            .build()
-        )
-
-        assert node.type == "div"
-        assert isinstance(node.content, list)
-        assert len(node.content) == 5
-        assert node.get_attribute("role") == "main"
-        assert "flex" in node.get_attribute("style")
-
-        # Test with mixed content types
-        mixed_content = ["Text", Node("br", None), "More text"]
-        mixed_node = NodeBuilder("p").content(mixed_content).build()
-
-        assert isinstance(mixed_node.content, list)
-        assert len(mixed_node.content) == 3
