--- tests/test_rendering_extended.py
@@ -1,590 +1,0 @@
-"""Extended Rendering Tests for Issue #491 Phase 4
-
-Extended tests for rendering system modules to boost coverage.
-Target: Rendering modules with partial coverage.
-"""
-
-from unittest.mock import Mock, patch
-
-import pytest
-
-from kumihan_formatter.core.ast_nodes import Node
-
-
-class TestRenderingSystemExtended:
-    """Extended tests for rendering system"""
-
-    def test_content_processor_complete(self):
-        """Test ContentProcessor complete functionality"""
-        try:
-            from kumihan_formatter.core.rendering.content_processor import (
-                ContentProcessor,
-            )
-
-            # Create processor with mock renderer
-            mock_renderer = Mock()
-            processor = ContentProcessor(mock_renderer)
-            assert processor is not None
-
-            # Test process_content with various inputs
-            test_cases = [
-                ("Simple text", str),
-                (["Text", "in", "list"], list),
-                (Node("p", "Node content"), object),
-                (None, type(None)),
-                ("", str),
-            ]
-
-            for content, expected_type in test_cases:
-                try:
-                    result = processor.process_content(content)
-                    if expected_type == str:
-                        assert isinstance(result, str)
-                except:
-                    pass
-
-            # Test recursive processing
-            nested_content = [
-                "Text",
-                Node("span", "Nested"),
-                ["Deep", Node("em", "nesting")],
-            ]
-
-            try:
-                result = processor.process_content(nested_content)
-                assert result is not None
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-    def test_html_escaping_complete(self):
-        """Test HTML escaping functionality"""
-        try:
-            from kumihan_formatter.core.rendering.html_escaping import (
-                escape_attribute,
-                escape_html,
-                unescape_html,
-            )
-
-            # Test HTML escaping
-            escape_tests = [
-                ("<script>alert('XSS')</script>", "&lt;script&gt;"),
-                ("Test & < > \" '", "Test &amp; &lt; &gt;"),
-                ("Normal text", "Normal text"),
-                ("", ""),
-                (None, None),
-            ]
-
-            for input_text, expected_partial in escape_tests:
-                if input_text is not None:
-                    try:
-                        result = escape_html(input_text)
-                        assert isinstance(result, str)
-                        if expected_partial:
-                            assert expected_partial in result
-                    except:
-                        pass
-
-            # Test attribute escaping
-            try:
-                result = escape_attribute('value with "quotes"')
-                assert '"' not in result or "&quot;" in result
-            except:
-                pass
-
-            # Test unescaping
-            try:
-                escaped = "&lt;p&gt;Test&lt;/p&gt;"
-                unescaped = unescape_html(escaped)
-                assert "<p>" in unescaped
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-    def test_compound_renderer_complete(self):
-        """Test CompoundElementRenderer completely"""
-        try:
-            from kumihan_formatter.core.rendering.compound_renderer import (
-                CompoundElementRenderer,
-            )
-
-            renderer = CompoundElementRenderer()
-            assert renderer is not None
-
-            # Test rendering compound elements
-            compound_types = [
-                "details",
-                "summary",
-                "blockquote",
-                "figure",
-                "figcaption",
-            ]
-
-            for elem_type in compound_types:
-                node = Node(elem_type, f"Content for {elem_type}")
-
-                try:
-                    method_name = f"render_{elem_type}"
-                    if hasattr(renderer, method_name):
-                        method = getattr(renderer, method_name)
-                        result = method(node)
-                        assert isinstance(result, str)
-                except:
-                    pass
-
-            # Test details/summary combination
-            try:
-                summary = Node("summary", "Click to expand")
-                content = Node("p", "Hidden content")
-                details = Node("details", [summary, content])
-
-                result = renderer.render_details(details)
-                assert isinstance(result, str)
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-    def test_list_renderer_complete(self):
-        """Test ListRenderer completely"""
-        try:
-            from kumihan_formatter.core.rendering.list_renderer import ListRenderer
-
-            renderer = ListRenderer()
-            assert renderer is not None
-
-            # Test unordered list
-            items = [Node("li", "Item 1"), Node("li", "Item 2"), Node("li", "Item 3")]
-            ul_node = Node("ul", items)
-
-            try:
-                result = renderer.render_ul(ul_node)
-                assert isinstance(result, str)
-                assert "<ul>" in result or result.startswith("ul")
-            except:
-                pass
-
-            # Test ordered list
-            ol_node = Node("ol", items)
-
-            try:
-                result = renderer.render_ol(ol_node)
-                assert isinstance(result, str)
-                assert "<ol>" in result or result.startswith("ol")
-            except:
-                pass
-
-            # Test list item
-            li_node = Node("li", "List item content")
-
-            try:
-                result = renderer.render_li(li_node)
-                assert isinstance(result, str)
-            except:
-                pass
-
-            # Test nested lists
-            nested_items = [
-                Node("li", "Parent item"),
-                Node(
-                    "li",
-                    [
-                        "Text before nested",
-                        Node("ul", [Node("li", "Nested 1"), Node("li", "Nested 2")]),
-                    ],
-                ),
-            ]
-            nested_list = Node("ul", nested_items)
-
-            try:
-                result = renderer.render_ul(nested_list)
-                assert isinstance(result, str)
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-    def test_div_renderer_complete(self):
-        """Test DivRenderer completely"""
-        try:
-            from kumihan_formatter.core.rendering.div_renderer import DivRenderer
-
-            renderer = DivRenderer()
-            assert renderer is not None
-
-            # Test basic div rendering
-            div_node = Node("div", "Basic div content")
-
-            try:
-                result = renderer.render_div(div_node)
-                assert isinstance(result, str)
-            except:
-                pass
-
-            # Test div with classes
-            class_tests = [
-                ("box", "content-box"),
-                ("highlight", "highlighted content"),
-                ("warning", "warning message"),
-                ("info", "info message"),
-                ("custom-class", "custom content"),
-            ]
-
-            for class_name, content in class_tests:
-                node = Node("div", content)
-                node.add_attribute("class", class_name)
-
-                try:
-                    result = renderer.render_div(node)
-                    assert isinstance(result, str)
-                except:
-                    pass
-
-            # Test div with multiple attributes
-            complex_div = Node("div", "Complex content")
-            complex_div.add_attribute("class", "container main")
-            complex_div.add_attribute("id", "main-content")
-            complex_div.add_attribute("data-value", "test")
-
-            try:
-                result = renderer.render_div(complex_div)
-                assert isinstance(result, str)
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-
-class TestBlockParserExtended:
-    """Extended tests for block parser"""
-
-    def test_block_parser_complete(self):
-        """Test BlockParser completely"""
-        try:
-            from kumihan_formatter.core.block_parser.block_parser import BlockParser
-
-            parser = BlockParser()
-            assert parser is not None
-
-            # Test parsing various block types
-            block_tests = [
-                (";;;太字;;; 内容 ;;;", "marker"),
-                ("> 引用文", "quote"),
-                ("```\ncode\n```", "code"),
-                ("---", "separator"),
-                ("# 見出し", "heading"),
-                ("普通のテキスト", "paragraph"),
-            ]
-
-            for content, expected_type in block_tests:
-                try:
-                    result = parser.parse_line(content)
-                    assert result is not None
-                except:
-                    pass
-
-            # Test multi-line parsing
-            multi_line_content = """# タイトル
-
-段落テキスト
-
-- リスト項目1
-- リスト項目2
-
-;;;強調;;; テキスト ;;;
-"""
-
-            try:
-                lines = multi_line_content.strip().split("\n")
-                results = []
-                for line in lines:
-                    result = parser.parse_line(line)
-                    if result:
-                        results.append(result)
-
-                assert len(results) > 0
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-    def test_block_validator(self):
-        """Test BlockValidator functionality"""
-        try:
-            from kumihan_formatter.core.block_parser.block_validator import (
-                BlockValidator,
-            )
-
-            validator = BlockValidator()
-            assert validator is not None
-
-            # Test validation methods
-            validation_tests = [
-                (";;;valid;;; content ;;;", True),
-                (";;;invalid", False),
-                ("((footnote))", True),
-                ("((incomplete", False),
-                ("normal text", True),
-            ]
-
-            for content, expected_valid in validation_tests:
-                try:
-                    result = validator.validate(content)
-                    assert isinstance(result, bool)
-                except:
-                    pass
-
-            # Test specific validators
-            try:
-                assert hasattr(validator, "validate_marker")
-                assert hasattr(validator, "validate_footnote")
-                assert hasattr(validator, "validate_ruby")
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-    def test_image_block_parser(self):
-        """Test ImageBlockParser functionality"""
-        try:
-            from kumihan_formatter.core.block_parser.image_block_parser import (
-                ImageBlockParser,
-            )
-
-            parser = ImageBlockParser()
-            assert parser is not None
-
-            # Test image syntax parsing
-            image_tests = [
-                "![alt text](image.jpg)",
-                "![](image.png)",
-                "![alt text](http://example.com/image.jpg)",
-                "![複雑な説明文](path/to/image.png)",
-            ]
-
-            for image_syntax in image_tests:
-                try:
-                    result = parser.parse(image_syntax)
-                    assert result is not None
-                    if hasattr(result, "type"):
-                        assert result.type in ["img", "image"]
-                except:
-                    pass
-
-            # Test image with attributes
-            try:
-                result = parser.parse('![alt](image.jpg){width="100" height="200"}')
-                assert result is not None
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-    def test_special_block_parser(self):
-        """Test SpecialBlockParser functionality"""
-        try:
-            from kumihan_formatter.core.block_parser.special_block_parser import (
-                SpecialBlockParser,
-            )
-
-            parser = SpecialBlockParser()
-            assert parser is not None
-
-            # Test special block types
-            special_blocks = [
-                ("```python\ncode\n```", "code"),
-                ("$$\nmath formula\n$$", "math"),
-                (":::\nnote content\n:::", "note"),
-                ("!!!\nwarning\n!!!", "warning"),
-            ]
-
-            for block_content, block_type in special_blocks:
-                try:
-                    lines = block_content.split("\n")
-                    result = parser.parse_block(lines)
-                    assert result is not None
-                except:
-                    pass
-
-        except ImportError:
-            pass
-
-
-class TestTOCGeneration:
-    """Test TOC (Table of Contents) generation"""
-
-    def test_toc_generator(self):
-        """Test TOC generator functionality"""
-        try:
-            from kumihan_formatter.core.toc_generator import TOCGenerator
-
-            generator = TOCGenerator()
-            assert generator is not None
-
-            # Create heading nodes
-            headings = [
-                Node("h1", "Chapter 1"),
-                Node("h2", "Section 1.1"),
-                Node("h2", "Section 1.2"),
-                Node("h3", "Subsection 1.2.1"),
-                Node("h1", "Chapter 2"),
-            ]
-
-            # Add IDs to headings
-            for i, heading in enumerate(headings):
-                heading.add_attribute("id", f"heading-{i}")
-
-            try:
-                toc = generator.generate(headings)
-                assert toc is not None
-                assert isinstance(toc, (str, list, dict))
-            except:
-                pass
-
-            # Test TOC formatting
-            try:
-                formatted = generator.format_toc(headings)
-                assert isinstance(formatted, str)
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-    def test_toc_formatter(self):
-        """Test TOC formatter functionality"""
-        try:
-            from kumihan_formatter.core.toc_formatter import TOCFormatter
-
-            formatter = TOCFormatter()
-            assert formatter is not None
-
-            # Test different TOC formats
-            toc_data = [
-                {"level": 1, "text": "Chapter 1", "id": "ch1"},
-                {"level": 2, "text": "Section 1.1", "id": "s11"},
-                {"level": 2, "text": "Section 1.2", "id": "s12"},
-            ]
-
-            # Test HTML format
-            try:
-                html_toc = formatter.to_html(toc_data)
-                assert isinstance(html_toc, str)
-                assert "<" in html_toc  # Should contain HTML tags
-            except:
-                pass
-
-            # Test markdown format
-            try:
-                md_toc = formatter.to_markdown(toc_data)
-                assert isinstance(md_toc, str)
-                assert "- " in md_toc or "* " in md_toc  # Markdown list
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-
-class TestHTMLFormatterExtended:
-    """Extended tests for HTML formatter"""
-
-    def test_html_formatter_complete(self):
-        """Test HTMLFormatter complete functionality"""
-        try:
-            from kumihan_formatter.core.rendering.html_formatter import HTMLFormatter
-
-            formatter = HTMLFormatter()
-            assert formatter is not None
-
-            # Test prettify functionality
-            ugly_html = "<div><p>Test</p><span>Content</span></div>"
-
-            try:
-                pretty = formatter.prettify(ugly_html)
-                assert isinstance(pretty, str)
-                assert len(pretty) >= len(ugly_html)  # Should add formatting
-            except:
-                pass
-
-            # Test minify functionality
-            try:
-                minified = formatter.minify(ugly_html)
-                assert isinstance(minified, str)
-            except:
-                pass
-
-            # Test validation
-            try:
-                is_valid = formatter.validate(ugly_html)
-                assert isinstance(is_valid, bool)
-            except:
-                pass
-
-            # Test fixing broken HTML
-            broken_html = "<div><p>Unclosed paragraph<div>Nested</div>"
-
-            try:
-                fixed = formatter.fix_html(broken_html)
-                assert isinstance(fixed, str)
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-    def test_html_tag_utils(self):
-        """Test HTML tag utilities"""
-        try:
-            from kumihan_formatter.core.rendering.html_tag_utils import (
-                close_tag,
-                create_tag,
-                self_closing_tag,
-                wrap_in_tag,
-            )
-
-            # Test create_tag
-            try:
-                tag = create_tag("div", {"class": "container", "id": "main"})
-                assert isinstance(tag, str)
-                assert "div" in tag
-                assert "class" in tag
-            except:
-                pass
-
-            # Test close_tag
-            try:
-                closing = close_tag("div")
-                assert closing == "</div>"
-            except:
-                pass
-
-            # Test self_closing_tag
-            try:
-                img = self_closing_tag("img", {"src": "test.jpg", "alt": "Test"})
-                assert isinstance(img, str)
-                assert "img" in img
-                assert "/>" in img or ">" in img
-            except:
-                pass
-
-            # Test wrap_in_tag
-            try:
-                wrapped = wrap_in_tag("p", "Content", {"class": "text"})
-                assert isinstance(wrapped, str)
-                assert "Content" in wrapped
-            except:
-                pass
-
-        except ImportError:
-            pass
