--- tests/test_utilities_complete.py
@@ -1,695 +1,0 @@
-"""Complete Utilities Tests for Issue #491 Phase 4
-
-Comprehensive tests for utility modules to boost coverage.
-Target: Various utility modules with low coverage.
-"""
-
-import json
-import tempfile
-from pathlib import Path
-from unittest.mock import MagicMock, Mock, patch
-
-import pytest
-
-
-class TestTextProcessorComplete:
-    """Complete tests for TextProcessor (currently 42% coverage)"""
-
-    def test_text_processor_all_methods(self):
-        """Test all TextProcessor methods"""
-        from kumihan_formatter.core.utilities.text_processor import TextProcessor
-
-        processor = TextProcessor()
-        assert processor is not None
-
-        # Test normalize_whitespace
-        test_cases = [
-            ("  hello   world  ", "hello world"),
-            ("\n\nhello\n\n", "hello"),
-            ("\t\ttab\t\t", "tab"),
-            ("multiple   spaces", "multiple spaces"),
-            ("", ""),
-            (None, None),
-        ]
-
-        for input_text, expected in test_cases:
-            if input_text is not None:
-                try:
-                    result = processor.normalize_whitespace(input_text)
-                    assert isinstance(result, str)
-                except:
-                    # Method may not exist
-                    pass
-
-        # Test clean_text
-        try:
-            result = processor.clean_text("  Test Text  ")
-            assert isinstance(result, str)
-        except:
-            pass
-
-        # Test process_text
-        try:
-            result = processor.process_text("Test\n\nText")
-            assert isinstance(result, str)
-        except:
-            pass
-
-        # Test strip methods
-        try:
-            result = processor.strip_html_tags("<p>Hello</p>")
-            assert "Hello" in result
-            assert "<p>" not in result
-        except:
-            pass
-
-        # Test escape methods
-        try:
-            result = processor.escape_special_chars("Test & < >")
-            assert isinstance(result, str)
-        except:
-            pass
-
-
-class TestLoggingUtilitiesComplete:
-    """Complete tests for logging utilities"""
-
-    def test_structured_logger_complete(self):
-        """Test StructuredLogger completely"""
-        from kumihan_formatter.core.utilities.structured_logger import StructuredLogger
-
-        logger = StructuredLogger()
-        assert logger is not None
-
-        # Test logging methods
-        log_methods = ["debug", "info", "warning", "error", "critical"]
-
-        for method_name in log_methods:
-            if hasattr(logger, method_name):
-                method = getattr(logger, method_name)
-                assert callable(method)
-
-                # Test basic logging
-                try:
-                    method("Test message")
-                except:
-                    pass
-
-                # Test structured logging
-                try:
-                    method({"message": "Test", "data": {"key": "value"}})
-                except:
-                    pass
-
-        # Test log formatting
-        try:
-            formatted = logger.format_log_entry(
-                "INFO", "Test message", {"extra": "data"}
-            )
-            assert isinstance(formatted, (str, dict))
-        except:
-            pass
-
-        # Test log filtering
-        try:
-            logger.set_level("INFO")
-            logger.add_filter(lambda record: record.get("level") != "DEBUG")
-        except:
-            pass
-
-    def test_logging_formatters(self):
-        """Test logging formatters"""
-        try:
-            from kumihan_formatter.core.utilities.logging_formatters import (
-                ColoredFormatter,
-                JSONFormatter,
-                PlainFormatter,
-            )
-
-            # Test JSON formatter
-            try:
-                json_formatter = JSONFormatter()
-                record = {"level": "INFO", "message": "Test"}
-                formatted = json_formatter.format(record)
-                assert isinstance(formatted, str)
-                # Should be valid JSON
-                parsed = json.loads(formatted)
-                assert isinstance(parsed, dict)
-            except:
-                pass
-
-            # Test plain formatter
-            try:
-                plain_formatter = PlainFormatter()
-                formatted = plain_formatter.format(record)
-                assert isinstance(formatted, str)
-            except:
-                pass
-
-            # Test colored formatter
-            try:
-                colored_formatter = ColoredFormatter()
-                formatted = colored_formatter.format(record)
-                assert isinstance(formatted, str)
-            except:
-                pass
-
-        except ImportError:
-            # Formatters may not be available
-            pass
-
-    def test_logging_handlers(self):
-        """Test logging handlers"""
-        try:
-            from kumihan_formatter.core.utilities.logging_handlers import (
-                ConsoleHandler,
-                FileHandler,
-                RotatingFileHandler,
-            )
-
-            # Test file handler
-            with tempfile.NamedTemporaryFile(
-                mode="w", suffix=".log", delete=False
-            ) as f:
-                temp_path = f.name
-
-            try:
-                file_handler = FileHandler(temp_path)
-                file_handler.emit({"level": "INFO", "message": "Test"})
-
-                # Check file was written
-                assert Path(temp_path).exists()
-            except:
-                pass
-            finally:
-                Path(temp_path).unlink(missing_ok=True)
-
-            # Test console handler
-            try:
-                console_handler = ConsoleHandler()
-                console_handler.emit({"level": "INFO", "message": "Test"})
-            except:
-                pass
-
-        except ImportError:
-            # Handlers may not be available
-            pass
-
-
-class TestFileSystemUtilities:
-    """Test file system utilities"""
-
-    def test_file_system_operations(self):
-        """Test file system operations"""
-        try:
-            from kumihan_formatter.core.utilities.file_system import (
-                copy_with_metadata,
-                ensure_directory,
-                get_file_info,
-                safe_remove,
-            )
-
-            with tempfile.TemporaryDirectory() as temp_dir:
-                temp_path = Path(temp_dir)
-
-                # Test ensure_directory
-                try:
-                    new_dir = temp_path / "test_dir"
-                    ensure_directory(new_dir)
-                    assert new_dir.exists()
-                    assert new_dir.is_dir()
-                except:
-                    pass
-
-                # Test get_file_info
-                test_file = temp_path / "test.txt"
-                test_file.write_text("Test content")
-
-                try:
-                    info = get_file_info(test_file)
-                    assert isinstance(info, dict)
-                    assert "size" in info
-                    assert "modified" in info
-                except:
-                    pass
-
-                # Test safe_remove
-                try:
-                    safe_remove(test_file)
-                    assert not test_file.exists()
-                except:
-                    pass
-
-        except ImportError:
-            # File system utilities may not be available
-            pass
-
-    def test_path_utilities(self):
-        """Test path utilities"""
-        try:
-            from kumihan_formatter.core.utilities.file_system import (
-                is_safe_path,
-                normalize_path,
-                resolve_path,
-            )
-
-            # Test normalize_path
-            try:
-                normalized = normalize_path("./test/../file.txt")
-                assert isinstance(normalized, (str, Path))
-            except:
-                pass
-
-            # Test resolve_path
-            try:
-                resolved = resolve_path("~/test.txt")
-                assert isinstance(resolved, (str, Path))
-            except:
-                pass
-
-            # Test is_safe_path
-            try:
-                assert is_safe_path("/tmp/test.txt") in [True, False]
-                assert is_safe_path("../../../etc/passwd") is False
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-
-class TestDataConverters:
-    """Test data converter utilities"""
-
-    def test_converters_functionality(self):
-        """Test converter functionality"""
-        try:
-            from kumihan_formatter.core.utilities.converters import (
-                from_json,
-                from_yaml,
-                to_dict,
-                to_json,
-                to_yaml,
-            )
-
-            # Test data
-            test_data = {
-                "name": "test",
-                "value": 123,
-                "nested": {"key": "value"},
-                "list": [1, 2, 3],
-            }
-
-            # Test to_json/from_json
-            try:
-                json_str = to_json(test_data)
-                assert isinstance(json_str, str)
-
-                parsed = from_json(json_str)
-                assert parsed == test_data
-            except:
-                pass
-
-            # Test to_dict
-            try:
-
-                class TestObj:
-                    def __init__(self):
-                        self.attr1 = "value1"
-                        self.attr2 = "value2"
-
-                obj = TestObj()
-                dict_result = to_dict(obj)
-                assert isinstance(dict_result, dict)
-                assert "attr1" in dict_result
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-    def test_type_converters(self):
-        """Test type converters"""
-        try:
-            from kumihan_formatter.core.utilities.converters import (
-                to_bool,
-                to_float,
-                to_int,
-                to_list,
-            )
-
-            # Test to_bool
-            bool_tests = [
-                ("true", True),
-                ("false", False),
-                ("1", True),
-                ("0", False),
-                ("yes", True),
-                ("no", False),
-            ]
-
-            for value, expected in bool_tests:
-                try:
-                    result = to_bool(value)
-                    assert result == expected
-                except:
-                    pass
-
-            # Test to_int
-            try:
-                assert to_int("123") == 123
-                assert to_int("123.45") == 123
-                assert to_int("invalid", default=0) == 0
-            except:
-                pass
-
-            # Test to_float
-            try:
-                assert to_float("123.45") == 123.45
-                assert to_float("123") == 123.0
-            except:
-                pass
-
-            # Test to_list
-            try:
-                assert to_list("a,b,c") == ["a", "b", "c"]
-                assert to_list([1, 2, 3]) == [1, 2, 3]
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-
-class TestPerformanceUtilities:
-    """Test performance utilities"""
-
-    def test_performance_decorators(self):
-        """Test performance decorators"""
-        try:
-            from kumihan_formatter.core.utilities.performance_decorators import (
-                cache,
-                profile,
-                retry,
-                timeit,
-            )
-
-            # Test timeit decorator
-            try:
-
-                @timeit
-                def test_func():
-                    return "result"
-
-                result = test_func()
-                assert result == "result"
-            except:
-                pass
-
-            # Test cache decorator
-            try:
-                call_count = 0
-
-                @cache
-                def cached_func(x):
-                    nonlocal call_count
-                    call_count += 1
-                    return x * 2
-
-                # First call
-                result1 = cached_func(5)
-                assert result1 == 10
-                assert call_count == 1
-
-                # Second call (should be cached)
-                result2 = cached_func(5)
-                assert result2 == 10
-                assert call_count == 1  # Should not increase
-            except:
-                pass
-
-            # Test retry decorator
-            try:
-                attempt_count = 0
-
-                @retry(max_attempts=3)
-                def flaky_func():
-                    nonlocal attempt_count
-                    attempt_count += 1
-                    if attempt_count < 2:
-                        raise Exception("Temporary error")
-                    return "success"
-
-                result = flaky_func()
-                assert result == "success"
-                assert attempt_count == 2
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-    def test_performance_trackers(self):
-        """Test performance trackers"""
-        try:
-            from kumihan_formatter.core.utilities.performance_trackers import (
-                MemoryTracker,
-                ResourceTracker,
-                TimeTracker,
-            )
-
-            # Test MemoryTracker
-            try:
-                mem_tracker = MemoryTracker()
-                mem_tracker.start()
-
-                # Do some memory allocation
-                data = [i for i in range(1000)]
-
-                usage = mem_tracker.get_usage()
-                assert isinstance(usage, (int, float))
-
-                mem_tracker.stop()
-                report = mem_tracker.get_report()
-                assert isinstance(report, dict)
-            except:
-                pass
-
-            # Test TimeTracker
-            try:
-                time_tracker = TimeTracker()
-                time_tracker.start("operation1")
-
-                # Simulate work
-                import time
-
-                time.sleep(0.01)
-
-                time_tracker.stop("operation1")
-                duration = time_tracker.get_duration("operation1")
-                assert isinstance(duration, float)
-                assert duration > 0
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-
-class TestStringUtilities:
-    """Test string manipulation utilities"""
-
-    def test_string_similarity(self):
-        """Test string similarity utilities"""
-        try:
-            from kumihan_formatter.core.utilities.string_similarity import (
-                fuzzy_match,
-                levenshtein_distance,
-                similarity_ratio,
-            )
-
-            # Test Levenshtein distance
-            try:
-                distance = levenshtein_distance("hello", "hallo")
-                assert distance == 1
-
-                distance = levenshtein_distance("test", "test")
-                assert distance == 0
-
-                distance = levenshtein_distance("", "test")
-                assert distance == 4
-            except:
-                pass
-
-            # Test similarity ratio
-            try:
-                ratio = similarity_ratio("hello", "hallo")
-                assert 0 <= ratio <= 1
-                assert ratio > 0.5  # Should be similar
-
-                ratio = similarity_ratio("test", "test")
-                assert ratio == 1.0
-            except:
-                pass
-
-            # Test fuzzy match
-            try:
-                assert fuzzy_match("hello", "hallo", threshold=0.8) is True
-                assert fuzzy_match("hello", "world", threshold=0.8) is False
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-    def test_string_formatting(self):
-        """Test string formatting utilities"""
-        try:
-            from kumihan_formatter.core.utilities.string_utils import (
-                dedent_text,
-                indent_text,
-                truncate,
-                wrap_text,
-            )
-
-            # Test truncate
-            try:
-                result = truncate("This is a long text", 10)
-                assert len(result) <= 10
-                assert result.endswith("...")
-            except:
-                pass
-
-            # Test wrap_text
-            try:
-                long_text = "This is a very long text that needs to be wrapped"
-                wrapped = wrap_text(long_text, width=20)
-                lines = wrapped.split("\n")
-                assert all(len(line) <= 20 for line in lines)
-            except:
-                pass
-
-            # Test indent_text
-            try:
-                text = "Line 1\nLine 2"
-                indented = indent_text(text, 4)
-                assert indented.startswith("    ")
-                assert "\n    " in indented
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-
-class TestDataStructures:
-    """Test data structure utilities"""
-
-    def test_data_structures(self):
-        """Test custom data structures"""
-        try:
-            from kumihan_formatter.core.utilities.data_structures import (
-                CaseInsensitiveDict,
-                FrozenDict,
-                OrderedSet,
-            )
-
-            # Test OrderedSet
-            try:
-                ordered_set = OrderedSet([3, 1, 4, 1, 5])
-                assert list(ordered_set) == [
-                    3,
-                    1,
-                    4,
-                    5,
-                ]  # Maintains order, removes duplicates
-
-                ordered_set.add(2)
-                assert 2 in ordered_set
-
-                ordered_set.remove(1)
-                assert 1 not in ordered_set
-            except:
-                pass
-
-            # Test CaseInsensitiveDict
-            try:
-                ci_dict = CaseInsensitiveDict()
-                ci_dict["Key"] = "value"
-
-                assert ci_dict["key"] == "value"
-                assert ci_dict["KEY"] == "value"
-                assert ci_dict["Key"] == "value"
-            except:
-                pass
-
-            # Test FrozenDict
-            try:
-                frozen = FrozenDict({"a": 1, "b": 2})
-
-                assert frozen["a"] == 1
-                assert frozen["b"] == 2
-
-                # Should not allow modification
-                with pytest.raises(Exception):
-                    frozen["c"] = 3
-            except:
-                pass
-
-        except ImportError:
-            pass
-
-
-class TestErrorHandlingUtilities:
-    """Test error handling utilities"""
-
-    def test_error_analyzer(self):
-        """Test error analyzer utilities"""
-        try:
-            from kumihan_formatter.core.utilities.error_analyzer import (
-                analyze_error,
-                get_error_context,
-                suggest_fix,
-            )
-
-            # Test error analysis
-            try:
-                error = ValueError("Invalid value: 'test'")
-                analysis = analyze_error(error)
-
-                assert isinstance(analysis, dict)
-                assert "type" in analysis
-                assert "message" in analysis
-                assert analysis["type"] == "ValueError"
-            except:
-                pass
-
-            # Test error context
-            try:
-
-                def problematic_function():
-                    x = 1
-                    y = 0
-                    return x / y
-
-                try:
-                    problematic_function()
-                except ZeroDivisionError as e:
-                    context = get_error_context(e)
-                    assert isinstance(context, dict)
-                    assert "traceback" in context
-            except:
-                pass
-
-            # Test fix suggestions
-            try:
-                error = FileNotFoundError("File 'test.txt' not found")
-                suggestions = suggest_fix(error)
-
-                assert isinstance(suggestions, list)
-                assert len(suggestions) > 0
-            except:
-                pass
-
-        except ImportError:
-            pass
